requirements.txt
git-ignore
playwright
python-dotenv
